backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
pydantic==2.9.2
PyYAML==6.0.2
python-dotenv==1.0.1